add setup.py to allow easy install and packaging
I'm planning on packaging this tool for Debian, and its much easier if
the project has a setup.py.  Also, users can do `pip install -e .` or
you could also publish this tool to pypi.org with this.

After merging this, it would be helpful to also have a tagged release
that matches the version in the setup.py.